fix: register pgvector type on connect so embeddings deserialize as float arrays
Embeddings were returned as strings (dtype '<U4707') from the sync
psycopg2 connection because pgvector's custom type was never registered.
Added an SQLAlchemy connect event listener in sync_session.py to call
register_vector() on each new dbapi connection.

Added a defensive np.array(..., dtype=np.float32) cast on the seed
embeddings used to compute the centroid, so the type is correct even if
pgvector isn't yet registered in a test or local context.

+++ worker/tasks/ml_inference.py
@@ -134,7 +134,7 @@ def _run_pgvector_recommender(payload: dict, db: Session) -> dict:
         if sid not in found_seed_ids:
             logger.warning("pgvector recommender: seed ID %r not in track_metadata — skipping", sid)
 
-    centroid = np.mean([np.array(r[1]) for r in rows], axis=0).tolist()
+    centroid = np.mean([np.array(r[1], dtype=np.float32) for r in rows], axis=0).tolist()
 
     candidates = db.execute(
         text(
